pyproject.toml
[project]
name = "file-transformer"
version = "0.1.0"
description = "MCP server for transforming and manipulating local files"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "PyPDF2>=3.0.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
file-transformer = "file_transformer.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/file_transformer"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
